# gwatop-backend/gwatop_openapi.yaml
openapi: 3.0.3
info:
  title: GwaTop API
  description: AI-powered academic scheduler for university students
  version: 1.0.0
servers:
  - url: https://api.gwatop.com/v1
    description: Production
  - url: http://localhost:8000/v1
    description: Local Development
security:
  - BearerAuth: []

# ─────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: 이메일 회원가입
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email:    { type: string, format: email, example: "mj@gwatop.com" }
                password: { type: string, minLength: 8, example: "password123" }
                name:     { type: string, example: "김민준" }
      responses:
        "201":
          description: 회원가입 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "409":
          description: 이미 존재하는 이메일
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/login:
    post:
      tags: [Auth]
      summary: 이메일 로그인
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: 로그인 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401":
          description: 인증 실패
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/social:
    post:
      tags: [Auth]
      summary: Apple / Google 소셜 로그인
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, id_token]
              properties:
                provider:  { type: string, enum: [apple, google] }
                id_token:  { type: string }
      responses:
        "200":
          description: 소셜 로그인 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Access Token 재발급
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:  { type: string }
                  expires_in:    { type: integer, example: 3600 }

  # ── SEMESTERS ───────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:   { type: string, example: "invalid_credentials" }
        message: { type: string, example: "이메일 또는 비밀번호가 올바르지 않습니다." }
    AuthResponse:
      type: object
      properties:
        access_token:   { type: string }
        refresh_token:  { type: string }
        expires_in:     { type: integer, example: 3600 }
        user:           { $ref: "#/components/schemas/User" }
    User:
      type: object
      properties:
        id:         { type: string, format: uuid }
        email:      { type: string, format: email }
        name:       { type: string }
        created_at: { type: string, format: date-time }
...